requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.8.0
Werkzeug==3.0.1

# ─── Database ─────────────────────────────────────────────────────────────────
PyMySQL==1.1.0
SQLAlchemy==2.0.49
cryptography==41.0.7

# ─── Machine Learning ─────────────────────────────────────────────────────────
scikit-learn==1.5.2
numpy==2.0.2
pandas==2.2.3
scipy==1.17.1
joblib==1.4.2

# ─── AI / LLM Providers ───────────────────────────────────────────────────────
groq==1.1.2
anthropic==0.95.0

# ─── Two-Factor Authentication ────────────────────────────────────────────────
pyotp==2.9.0
qrcode==8.2
pillow==12.2.0

# ─── Background Scheduler ─────────────────────────────────────────────────────
APScheduler==3.10.4
tzlocal==5.3.1

# ─── HTTP & Networking ────────────────────────────────────────────────────────
requests==2.31.0
httpx==0.28.1
certifi==2026.2.25

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
graphviz==0.20.3
